+++ src/main/typescript/config.ts
@@ -4,6 +4,7 @@ import WvActorSheet from "./applications/actor/wvActorSheet.js";
 import EffectSheet from "./applications/item/effectSheet.js";
 import WeaponSheet from "./applications/item/weaponSheet.js";
 import { CONSTANTS, TYPES } from "./constants.js";
+import { CHARACTER_JSON_SCHEMA } from "./data/actor/character/source.js";
 import { BASE_ITEM_JSON_SCHEMA } from "./data/item/baseItem.js";
 import { WEAPON_SOURCE_JSON_SCHEMA } from "./data/item/weapon/source.js";
 import { getGame } from "./foundryHelpers.js";
@@ -26,6 +27,9 @@ export function configureFoundryOnInit(): void {
     ajv,
     macros,
     validators: {
+      actor: {
+        character: ajv.compile(CHARACTER_JSON_SCHEMA)
+      },
       item: {
         effect: ajv.compile(BASE_ITEM_JSON_SCHEMA),
         weapon: ajv.compile(WEAPON_SOURCE_JSON_SCHEMA)