+++ src/main/typescript/data/actor/character/leveling/properties.ts
@@ -1,0 +1,10 @@
+import LevelingSource from "./source.js";
+
+/** Derived leveling related data */
+export default class Leveling extends LevelingSource {
+  /** The current level of an Actor */
+  level?: number;
+
+  /** The maximum skill points of an Actor */
+  maxSkillPoints?: number;
+}