+++ src/main/typescript/data/actor/character/vitals/properties.ts
@@ -1,0 +1,7 @@
+import VitalsSource from "./source.js";
+
+/** Derived vitals data */
+export default class Vitals extends VitalsSource {
+  /** The healing rate of an Actor per 8 hours of rest */
+  healingRate?: number;
+}